+++ packages/strapi-admin/admin/src/translations/en.json
@@ -250,6 +250,7 @@
   "app.utils.add-filter": "Add filter",
   "app.utils.defaultMessage": " ",
   "app.utils.delete": "Delete",
+  "app.utils.delete-entry": "Delete this entry",
   "app.utils.publish": "Publish",
   "app.utils.filters": "Filters",
   "app.static.links.cheatsheet": "CheatSheet",
