+++ packages/strapi-plugin-content-manager/admin/src/containers/EditView/Header.js
@@ -3,22 +3,16 @@ import { useIntl } from 'react-intl';
 import { Header as PluginHeader } from '@buffetjs/custom';
 import { get, isEqual, isEmpty, toString } from 'lodash';
 
-import { PopUpWarning, request, templateObject, useGlobalContext } from 'strapi-helper-plugin';
+import { PopUpWarning, templateObject } from 'strapi-helper-plugin';
 
 import pluginId from '../../pluginId';
 import useDataManager from '../../hooks/useDataManager';
 import useEditView from '../../hooks/useEditView';
 
-const getRequestUrl = path => `/${pluginId}/explorer/${path}`;
-
 const Header = () => {
   const [showWarningCancel, setWarningCancel] = useState(false);
-  const [showWarningDelete, setWarningDelete] = useState(false);
-  const [didDeleteEntry, setDidDeleteEntry] = useState(false);
-  const [isModalConfirmButtonLoading, setIsModalConfirmButtonLoading] = useState(false);
   const { formatMessage } = useIntl();
   const formatMessageRef = useRef(formatMessage);
-  const { emitEvent } = useGlobalContext();
   const {
     initialData,
     isCreatingEntry,
@@ -28,10 +22,7 @@ const Header = () => {
     layout,
     modifiedData,
     onPublish,
-    redirectToPreviousPage,
     resetData,
-    slug,
-    clearData,
   } = useDataManager();
   const {
     allowedActions: { canDelete, canUpdate, canCreate, canPublish },
@@ -101,24 +92,6 @@ const Header = () => {
       });
     }
 
-    if (!isCreatingEntry && canDelete) {
-      headerActions.unshift({
-        label: formatMessage({
-          id: 'app.utils.delete',
-        }),
-        color: 'delete',
-        onClick: () => {
-          toggleWarningDelete();
-        },
-        type: 'button',
-        style: {
-          paddingLeft: 15,
-          paddingRight: 15,
-          fontWeight: 600,
-        },
-      });
-    }
-
     return headerActions;
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [
@@ -145,56 +118,10 @@ const Header = () => {
   }, [headerActions, headerTitle, apiID]);
 
   const toggleWarningCancel = () => setWarningCancel(prevState => !prevState);
-  const toggleWarningDelete = () => setWarningDelete(prevState => !prevState);
 
   const handleConfirmReset = () => {
     toggleWarningCancel();
     resetData();
-  };
-
-  const handleConfirmDelete = async () => {
-    try {
-      // Show the loading state
-      setIsModalConfirmButtonLoading(true);
-
-      emitEvent('willDeleteEntry');
-
-      await request(getRequestUrl(`${slug}/${initialData.id}`), {
-        method: 'DELETE',
-      });
-
-      strapi.notification.success(`${pluginId}.success.record.delete`);
-
-      emitEvent('didDeleteEntry');
-
-      // This is used to perform action after the modal is closed
-      // so the transitions are smoother
-      // Actions will be performed in the handleClosed function
-      setDidDeleteEntry(true);
-    } catch (err) {
-      emitEvent('didNotDeleteEntry', { error: err });
-      const errorMessage = get(
-        err,
-        'response.payload.message',
-        formatMessage({ id: `${pluginId}.error.record.delete` })
-      );
-      strapi.notification.error(errorMessage);
-    } finally {
-      setIsModalConfirmButtonLoading(false);
-      toggleWarningDelete();
-    }
-  };
-
-  const handleClosed = () => {
-    if (didDeleteEntry) {
-      if (!isSingleType) {
-        redirectToPreviousPage();
-      } else {
-        clearData();
-      }
-    }
-
-    setDidDeleteEntry(false);
   };
 
   return (
@@ -211,20 +138,6 @@ const Header = () => {
         }}
         popUpWarningType="danger"
         onConfirm={handleConfirmReset}
-      />
-      <PopUpWarning
-        isOpen={showWarningDelete}
-        toggleModal={toggleWarningDelete}
-        content={{
-          title: `${pluginId}.popUpWarning.title`,
-          message: `${pluginId}.popUpWarning.bodyMessage.contentType.delete`,
-          cancel: `${pluginId}.popUpWarning.button.cancel`,
-          confirm: `${pluginId}.popUpWarning.button.confirm`,
-        }}
-        popUpWarningType="danger"
-        onConfirm={handleConfirmDelete}
-        onClosed={handleClosed}
-        isConfirmButtonLoading={isModalConfirmButtonLoading}
       />
     </>
   );
